requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
gunicorn==25.1.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==1.23.5
packaging==26.0
pillow==10.2.0
psycopg2-binary==2.9.11
pydicom==3.0.1
Werkzeug==3.1.5
sendgrid
reportlab
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.15.0
gdown